101-print_number.c

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -6,9 +6,9 @@
  */
 void print_number(int n)
 {
-long m; /* power of 10 */
-int c; /* boolean check */
-long num; /* convert int to long */
+int m;
+int c;
+int num;
 num n;
 if (num < 0)
 {
